+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Version `v0.24.0`
 
+* ![Feature][badge-feature] Deployment is now more customizable and thus not as tied to Travis CI as before. ([#1147][github-1147])
+
 * ![Enhancement][badge-enhancement] In the PDF/LaTeX output, images that are wider than the text are now being scaled down to text width automatically. The PDF builds now require the [adjustbox](https://ctan.org/pkg/adjustbox) LaTeX package to be available. ([#1137][github-1137])
 
 ## Version `v0.23.3`
@@ -412,6 +414,7 @@
 [github-1082]: https://github.com/JuliaDocs/Documenter.jl/pull/1082
 [github-1115]: https://github.com/JuliaDocs/Documenter.jl/pull/1115
 [github-1137]: https://github.com/JuliaDocs/Documenter.jl/pull/1137
+[github-1147]: https://github.com/JuliaDocs/Documenter.jl/pull/1147
 
 [documenterlatex]: https://github.com/JuliaDocs/DocumenterLaTeX.jl
 [documentermarkdown]: https://github.com/JuliaDocs/DocumenterMarkdown.jl